refactor: ToBuyItem을 ToBuyItemMain, ToBuyItemOptions로 분리
